moved emitted event to types crate
