+++ crates/types/src/lib.rs
@@ -3,6 +3,7 @@ pub mod contract_class;
 pub mod contract_storage_key;
 pub mod error;
 pub mod felt;
+pub mod models;
 pub mod patricia_key;
 pub mod serde_helpers;
 pub mod traits;
